Add basic reflection API

+++ Examples/Base/Include/Examples-Base/ImGuiRenderer.hpp
@@ -11,7 +11,7 @@
 struct ImGuiRendererCreateInfo
 {
     RHI::Context* context;
-    RHI::TL::Vector<uint8_t> shaderBlob;
+    RHI::TL::Vector<uint32_t> shaderBlob;
     RHI::CommandPool* commandAllocator;
 };
